Refactoring printer_local_comm to be consistent with TSDApp
